refactor: promote ManaSymbolRichCtrl v2 into the canonical module
The v2 rewrite (overlay-based hint) is confirmed working. Drop v1,
rename v2 into widgets/mana_rich_text_ctrl.py, and point the three
callers back at the canonical path.

Also runs the standard ruff + black sweep on the changed files:
  - drop the forward-reference string on the ManaIconFactory annotation
    (UP037; redundant under `from __future__ import annotations`)
  - re-sort the imports in widgets/app_frame.py (I001)
  - reformat utils/constants/keyboard.py and the new ctrl file to match
    the project's line-length/style conventions.
