+++ widgets/panels/card_inspector_panel.py
@@ -33,7 +33,7 @@ from utils.constants import (
 from utils.mana_icon_factory import ManaIconFactory
 from utils.stylize import stylize_button, stylize_textctrl
 from widgets.card_image_display import CardImageDisplay
-from widgets.mana_rich_text_ctrl_2 import ManaSymbolRichCtrl
+from widgets.mana_rich_text_ctrl import ManaSymbolRichCtrl
 
 
 class CardInspectorPanel(wx.Panel):
